First working prototyp of dynamic DNS server
use the web http form api to register an ip with a domain and then query
the dns entry for it. Remember to use the fqdn (with the dot at the end).

At the moment everything crashes if you do a dns query for a non existant
domain but it seemed like a good point to commit